move connection objects to connection folder
