+++ src/model/networking/connection/client-connection.cc
